BE/requirements.txt
alembic==1.13.1
ultralytics==8.0.196
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==5.0.0
billiard==4.2.2
CacheControl==0.14.3
cachetools==6.2.1
celery==5.3.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.109.0
filelock==3.20.0
firebase-admin==6.4.0
flatbuffers==25.9.23
fsspec==2025.9.0
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.26.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numpy==1.26.4
onnx==1.15.0
onnxruntime==1.16.3
opencv-python==4.9.0.80
packaging==25.0
passlib==1.7.4
pillow==10.2.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.10.1
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
redis==5.0.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
sympy==1.14.0
torch==2.1.2
torchvision==0.16.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.27.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
